requirements.txt
Django==2.2.5
djangorestframework==3.10.3
Pillow==6.1.0
psycopg2-binary==2.8.3
pypi==2.1
pytz==2019.2
sqlparse==0.3.0
